update scss better background
