+++ src/app/app.component.html
@@ -1,7 +1,7 @@
 
+<div class="background"></div>
 <div class="container">
   <div *ngIf="!isGameStarted()">
-
     <label for="playersNumber1">Players number:</label>
     <select name="playersNumber" id="playersNumber1" [(ngModel)]="playersNumber.value">
       <option disabled>How many player?</option>
@@ -11,12 +11,19 @@
   </div>
 
   <div class="game-container" *ngIf="isGameStarted()">
-    <label for="playersNumber">Players number:</label>
-    <select name="playersNumber" id="playersNumber" [(ngModel)]="playersNumber.value">
-      <option disabled>How many player?</option>
-      <option *ngFor="let playerNumber of selectModel" [value]="playerNumber.value">{{playerNumber.title}}</option>
-    </select>
-    <button *ngIf="isGameStarted()" class="btn btn-round btn-filled-green btn-md" (click)="onStartClick()">Restart Game</button>
+
+    <fa-icon class="menu-btn" [icon]="['fas', 'bars']" (click)="openGameMenu()"></fa-icon>
+    <div *ngIf="gameMenuOpened" class="game-menu">
+      <form>
+        <label for="playersNumber">Players number:</label>
+        <select name="playersNumber" id="playersNumber" [(ngModel)]="playersNumber.value">
+          <option disabled>How many player?</option>
+          <option *ngFor="let playerNumber of selectModel" [value]="playerNumber.value">{{playerNumber.title}}</option>
+        </select>
+      </form>
+      <button *ngIf="isGameStarted()" class="btn btn-round btn-filled-green btn-md" (click)="onStartClick()">Restart Game</button>
+    </div>
+
     <div class="table">
       <div class="deck"></div>
       <!-- <div class="discard-pile"></div> -->
@@ -24,7 +31,7 @@
     <div *ngFor="let hand of hands|keyvalue;let pIndex=index;" class="player player{{pIndex}}">
       <div class="hand">
         <div class="deck">
-          <app-card *ngFor="let card of hands[pIndex];let index=index;" [card]="card" [hidden]="true"
+          <app-card *ngFor="let card of hands[pIndex];let index=index;" [card]="card"
             class="card-{{index}}"></app-card>
         </div>
       </div>
