+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { Card, symbolsTable } from "./model/card";
+import { Card } from "./model/card";
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
@@ -7,6 +7,7 @@ import { Card, symbolsTable } from "./model/card";
 })
 export class AppComponent {
   title = 'daifuqo';
+  gameMenuOpened:boolean =false;
   deck: any[] = [];
   hands: Array<Array<Card>> = [];
   gameState:string = 'stopped';
@@ -17,7 +18,9 @@ export class AppComponent {
     { title: "5 players", value: 5, },
     { title: "6 players", value: 6, },
   ]
-
+  openGameMenu(){
+    this.gameMenuOpened = true;
+  }
   initializeDeck() {
     let neutralCards = [2, 3, 4, 5, 6, 7, 8, 9, 10, 'J', 'Q', 'K', 'A'];
     let cardsOfSpades = neutralCards.map(powers => new Card(powers, "spades"));
@@ -75,75 +78,9 @@ export class AppComponent {
     console.log("Game started",this.playersNumber);
     this.gameState ='started';
     this.initializeDeck();
-    console.log("Normal deck",this.deck);
     this.deck = this.shuffle([...this.deck]);
-    console.log("Shuffle deck",this.deck);
     this.giveCards();
     console.log(this.hands);
-  }
-
-
-  getCardFullName($card: Card) {
-    switch ($card.getCardPower()) {
-      case 'A': return `a${$card.getType()}.svg`;
-      case 'K': return 'king.svg';
-      case 'Q': return 'queen.svg';
-      case 'J': return 'jota.svg';
-      case 'Joker': return 'joker.svg';
-      default: return '';
-    }
-  }
-
-  createRange(number: number): Array<number> {
-    return new Array(number);
-  }
-  isRotated(item: number, $card: Card): boolean {
-    let name: any = $card.getCardPower();
-    return item === 3 && [10, 9].includes(name)
-  }
-
-
-  isBig($card: Card) {
-    let name: any = $card.getCardPower();
-    return [10, 8].includes(name)
-  }
-
-  isHuge($card: Card) {
-    let name: any = $card.getCardPower();
-    return [7].includes(name)
-  }
-
-  isInnerCentered($card: Card) {
-    let name = $card.getCardPower();
-    return [2, 3, 'A', 'J', 'Q', 'K'].includes(name);
-  }
-  isKQJA($card: Card) {
-    let name: any = $card.getCardPower();
-    return ['A', 'J', 'Q', 'K'].includes(name);
-  }
-  convertStringToNumber(input: string | undefined): number {
-    if (!input) return NaN;
-    if (typeof input === 'number') {
-      return input
-    };
-    if (input.trim().length == 0) {
-      return NaN;
-    }
-    return Number(input);
-  }
-  getSymbolsTable($card: Card): { col_up: any; col_center: number; col_down: any; } {
-    let name: any = $card.getCardPower()
-    let value = this.convertStringToNumber(['A', 'J', 'K', 'Q'].includes(name) ? '0' : name);
-    return symbolsTable[value ? value - 1 : 0];
-  }
-
-  getColumnUp($card: Card) {
-    return this.getSymbolsTable($card).col_up;
-  }
-  getColumnCenter($card: Card) {
-    return this.getSymbolsTable($card).col_center;
-  }
-  getColumnDown($card: Card) {
-    return this.getSymbolsTable($card).col_down;
+    this.gameMenuOpened=false;
   }
 }
