+++ src/app/app.module.ts
@@ -4,6 +4,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 @NgModule({
     declarations: [
@@ -15,6 +16,7 @@ import { SharedModule } from './shared/shared.module';
         BrowserModule,
         AppRoutingModule,
         SharedModule,
+        FontAwesomeModule,
     ]
 })
 export class AppModule { }
