+++ src/app/model/Card.ts
@@ -8,7 +8,61 @@ export class Card {
 
     getCardPower(): number | string { return this.cardPower; }
     getType():  string { return this.type; }
+    convertStringToNumber(input: string | undefined): number {
+      if (!input) return NaN;
+      if (typeof input === 'number') {
+        return input
+      };
+      if (input.trim().length == 0) {
+        return NaN;
+      }
+      return Number(input);
+    }
+    getSymbolsTable(): { col_up: any; col_center: number; col_down: any; } {
+      let name: any = this.getCardPower()
+      let value = this.convertStringToNumber(['A', 'J', 'K', 'Q'].includes(name) ? '0' : name);
+      return symbolsTable[value ? value - 1 : 0];
+    }
+    getColumnUp() {
+      return this.getSymbolsTable().col_up;
+    }
+    getColumnCenter() {
+      return this.getSymbolsTable().col_center;
+    }
+    getColumnDown() {
+      return this.getSymbolsTable().col_down;
+    }
+    getCardFullName() {
+      switch (this.getCardPower()) {
+        case 'A': return `a${this.getType()}.svg`;
+        case 'K': return 'king.svg';
+        case 'Q': return 'queen.svg';
+        case 'J': return 'jota.svg';
+        case 'Joker': return 'joker.svg';
+        default: return '';
+      }
+    }
+    isBig() {
+      let name: any = this.getCardPower();
+      return [10, 8].includes(name)
+    }
 
+    isHuge() {
+      let name: any = this.getCardPower();
+      return [7].includes(name)
+    }
+    isRotated(item: number): boolean {
+      let name: any = this.getCardPower();
+      return item === 3 && [10, 9].includes(name)
+    }
+    isInnerCentered() {
+      let name = this.getCardPower();
+      return [2, 3, 'A', 'J', 'Q', 'K'].includes(name);
+    }
+    isKQJA() {
+      let name: any = this.getCardPower();
+      return ['A', 'J', 'Q', 'K'].includes(name);
+    }
     isGreaterThan($card: Card, $isRevolution = false): boolean {
       let givenCardPower: string | number = $card.getCardPower();
       if (givenCardPower === this.cardPower) {
